Rework-Audit 1a Schritt 2, 1e, Teil 6: Backends migriert, Editor-Widgets, Content
1a Schritt 2 — alle fuenf Backends nutzen jetzt die gemeinsame Schicht
(SsaoKernel, SkyNoise3D, SkyFrameParams/CloudWindVector, LightPacking,
ClipSpace, RenderConstants, RenderSorter) statt eigener Kopien. Dazu die
~850 identischen HLSL-Zeilen aus D3D11/D3D12 in D3D_Shared/HlslSources.h.
Damit werden die beiden beim Diffen gefundenen Bugs wirksam:
- Wolken-Wind driftete auf Vulkan/D3D11/D3D12 um 180 Grad falsch (fehlende
  Z-Negation gegenueber der GL/Metal-Referenz).
- Vulkan/D3D klassifizierten Transparenz ohne instanceTint.a, ein per Alpha
  ausblendender Partikel landete im Opaque-Pass.
GL: der 8-Licht-Block inkl. CSM-Binding war dreimal in DrawScene kopiert ->
BindSceneLighting; m_lastInvViewProj/m_lastSunDir waren write-only, geloescht.
Metal: die toten PoC-Pfade EnsureShadercDemoPipeline/EnsureShadercTestMesh
(~350 Zeilen) entfernt, die zwei SkyParams-Fuellungen auf einen Aufruf.
Vulkan: der Shader-Drift zu den GL-eingebetteten Quellen ist jetzt im Kopf von
sky.frag/scene.frag dokumentiert statt stillschweigend.
ShaderManager.h exportiert kein globales `namespace fs` mehr in jeden Consumer.

1e — assetDropSlot ersetzt ~20 kopierte HE_ASSET_PATH-Drop-Slots und
vereinheitlicht dabei die Undo-Snapshots, die in den Graph-Panels fehlten;
AssetPanelState<T> loest die 9-fache Panel-State-Boilerplate ab (auf dem
bereits vorhandenen EditorAssetTypeCache aufgesetzt, keine zweite Cache-Ebene).
HcClassList heisst jetzt HcEditorUtil.

Teil 6 (ContentManager/Hpak) — lookupAssetMutable-Template, Shader-Variant-
Codec als Template, readStoredEntry/readEntry zusammengelegt, JSON-Escaper
durch nlohmann ersetzt (mit demselben replace-Error-Handler, sonst waere der
UTF-8-Crash zurueck), __scene_index__ als Konstante, kFlagUsesDict wird jetzt
explizit abgelehnt statt still ignoriert, pushInstanceMethod fuer die fuenf
Lua-Call-Kopien.

Nachweis: 902/902 Tests gruen. Zusaetzlich Metal auf echter Hardware
verifiziert — der zeitanimierte Wolken-Frame ist zwischen zwei Laeufen
desselben Builds zu 48,8 % verschieden, taugt also nicht zum Vergleich; mit
abgeschalteten Wolken ist der Frame deterministisch und vor/nach der Migration
BYTE-IDENTISCH.

+++ src/HE_Core/include/ContentManager/Assets.h
@@ -328,9 +328,13 @@ struct TextureAsset : public RuntimeAsset
 	// 0 as the leading width*height*channels bytes, so appended mips are
 	// backward-compatible (ignored by consumers that don't sample them).
 	std::vector<uint8_t> data;     // raw pixel bytes (RGBA8 unless channels says otherwise)
-	size_t               width    = 0;
-	size_t               height   = 0;
-	size_t               channels = 0;
+	// Fixed-width on purpose: these go into the TXMI chunk verbatim, and as size_t
+	// the on-disk layout differed between 32- and 64-bit builds (a pak written by
+	// one was unreadable by the other). See HAsset::readTextureHeader, which still
+	// reads the old 64-bit layout.
+	uint32_t             width    = 0;
+	uint32_t             height   = 0;
+	uint32_t             channels = 0;
 
 	// Pack-time cook metadata (CHUNK_TXMI tail; absent → the defaults below, so
 	// old assets keep loading unchanged).
